Switch to using git branch pyk version
